Fix Android NDK/clang build and SDL_main entry point
Make the codebase build with the Android NDK (clang) and fix the native
entry point that prevented the APK from launching.

Build configuration:
- Build the game as a shared library named "main" on Android instead of
  an executable, since SDL3 loads the game via libmain.so
- Link the prebuilt arm64 libbass.so; it was packaged through jniLibs but
  never linked, so every BASS_* symbol was undefined
- Only define SEXY_USE_OPENGL off Android; desktop OpenGL is unavailable
  there and pulled in GL 3.x core symbols that libGLES does not export
- Skip copying the Windows DLLs from thirdparty/ on Android

Entry point:
- Include SDL3/SDL_main.h on Android so main() becomes SDL_main, which is
  the symbol SDL's loader looks up. This fixes the "Couldn't find
  function SDL_main in library libmain.so" crash on startup

Portability:
- Provide stricmp/strnicmp on non-Windows via strcasecmp/strncasecmp
- Add a timeGetTime() shim based on SDL_GetTicks()
- Give forward-declared enums a fixed underlying type (: int)
- Add missing <thread>, <list> and <cmath> includes
- Replace _localtime64_s with localtime_r and 0i64 with 0LL
- Silence -Wformat-security by using fputs instead of printf/fprintf
- Guard Win32-only code (ChangeDisplayWindowEnumProc, SetThreadPriority,
  ReleaseCapture, -recnum demo options) behind _WIN32

+++ src/FrameworkResources.cpp
@@ -1,5 +1,6 @@
 #include "FrameworkResources.h"
 #include "SexyAppFramework/ResourceManager.h"
+#include <cstdint>
 using namespace Sexy;
 
 #pragma warning(disable:4311 4312)
@@ -118,7 +119,7 @@ static Sexy::ResoddedFrameworkResourceId ResoddedFrameworkGetIdByVariable(const 
 			aMap[*(int*)gResources[i]] = i;
 	}
 
-	MyMap::iterator anItr = aMap.find((int)theVariable);
+	MyMap::iterator anItr = aMap.find((int)(intptr_t)theVariable);
 	if (anItr == aMap.end())
 		return ResoddedFrameworkRESOURCE_ID_MAX;
 	else
